+++ Source/GTT/Public/Vehicles/GTTFarmVanPawn.h
@@ -1,0 +1,35 @@
+#pragma once
+
+#include "CoreMinimal.h"
+#include "Vehicles/GTTVehicleBase.h"
+#include "GTTFarmVanPawn.generated.h"
+
+class UStaticMeshComponent;
+
+UCLASS()
+class GTT_API AGTTFarmVanPawn : public AGTTVehicleBase
+{
+    GENERATED_BODY()
+
+public:
+    AGTTFarmVanPawn();
+
+protected:
+    UPROPERTY(VisibleAnywhere, BlueprintReadOnly, Category="GTT|Vehicle|FarmVan")
+    TObjectPtr<UStaticMeshComponent> CabinMesh;
+
+    UPROPERTY(VisibleAnywhere, BlueprintReadOnly, Category="GTT|Vehicle|FarmVan")
+    TObjectPtr<UStaticMeshComponent> CargoMesh;
+
+    UPROPERTY(VisibleAnywhere, BlueprintReadOnly, Category="GTT|Vehicle|FarmVan")
+    TObjectPtr<UStaticMeshComponent> LeftFrontWheel;
+
+    UPROPERTY(VisibleAnywhere, BlueprintReadOnly, Category="GTT|Vehicle|FarmVan")
+    TObjectPtr<UStaticMeshComponent> RightFrontWheel;
+
+    UPROPERTY(VisibleAnywhere, BlueprintReadOnly, Category="GTT|Vehicle|FarmVan")
+    TObjectPtr<UStaticMeshComponent> LeftRearWheel;
+
+    UPROPERTY(VisibleAnywhere, BlueprintReadOnly, Category="GTT|Vehicle|FarmVan")
+    TObjectPtr<UStaticMeshComponent> RightRearWheel;
+};